+++ common/src/main/java/com/lowdragmc/lowdraglib/gui/util/DrawerHelper.java
@@ -20,6 +20,7 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
 import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.network.chat.Component;
@@ -401,6 +402,11 @@ public class DrawerHelper {
 
     @ExpectPlatform
     public static void drawTooltip(GuiGraphics graphics, int mouseX, int mouseY, List<Component> tooltipTexts, ItemStack tooltipStack, @Nullable TooltipComponent tooltipComponent, Font tooltipFont) {
+        throw new AssertionError();
+    }
+
+    @ExpectPlatform
+    public static ClientTooltipComponent getClientTooltipComponent(TooltipComponent component) {
         throw new AssertionError();
     }
 }