+++ fabric/src/main/java/com/lowdragmc/lowdraglib/gui/util/fabric/DrawerHelperImpl.java
@@ -2,6 +2,8 @@ package com.lowdragmc.lowdraglib.gui.util.fabric;
 
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
+import net.fabricmc.fabric.api.client.rendering.v1.TooltipComponentCallback;
+import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.inventory.tooltip.TooltipComponent;
 import net.minecraft.world.item.ItemStack;
@@ -18,4 +20,13 @@ public class DrawerHelperImpl {
     public static void drawTooltip(GuiGraphics graphics, int mouseX, int mouseY, List<Component> tooltipTexts, ItemStack tooltipStack, TooltipComponent tooltipComponent, Font tooltipFont) {
         graphics.renderTooltip(tooltipFont, tooltipTexts, Optional.ofNullable(tooltipComponent), mouseX, mouseY);
     }
+
+    public static ClientTooltipComponent getClientTooltipComponent(TooltipComponent component) {
+        var result = TooltipComponentCallback.EVENT.invoker().getComponent(component);
+        if (result == null) {
+            result = ClientTooltipComponent.create(component);
+        }
+        return result;
+    }
+
 }