+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # ChangeLog
 
-* compass system
+* fix widget rendering issue
+* fix renderer is gui 3d
+* fix widget hotswap
+* fix emi tooltips